+++ hooks/useOutsideClick.tsx
@@ -1,0 +1,19 @@
+import { useEffect } from "react";
+
+const useOutsideClick = (ref: React.RefObject<HTMLElement>, callback: () => void) => {
+  useEffect(() => {
+    const handleOutsideClick = (event: MouseEvent) => {
+      if (ref.current && !ref.current.contains(event.target as Node)) {
+        callback();
+      }
+    };
+
+    document.addEventListener("mousedown", handleOutsideClick);
+
+    return () => {
+      document.removeEventListener("mousedown", handleOutsideClick);
+    };
+  }, [ref, callback]);
+};
+
+export default useOutsideClick;
